+++ src/main/java/net/jrodolfo/java_evolution/java01/javabeans/package-info.java
@@ -1,4 +1,4 @@
 /**
- * Explanatory module for JavaBeans.
+ * Executable examples for JavaBeans.
  */
 package net.jrodolfo.java_evolution.java01.javabeans;