+++ src/main/java/net/jrodolfo/java_evolution/java02/java2d/package-info.java
@@ -1,4 +1,4 @@
 /**
- * Explanatory module for Java 2D.
+ * Executable examples for Java 2D.
  */
 package net.jrodolfo.java_evolution.java02.java2d;