+++ src/main/java/net/jrodolfo/java_evolution/java06/monitoring_management/package-info.java
@@ -1,4 +1,4 @@
 /**
- * Explanatory module for Java 6 monitoring and management improvements.
+ * Executable examples for Java 6 monitoring and management improvements.
  */
 package net.jrodolfo.java_evolution.java06.monitoring_management;